requirements.txt
pandas
scikit-learn
customtkinter
seaborn
customtkinter
matplotlib
numpy
pandas
pillow
scikit-learn
seaborn